build: add action to publish with provenance
We added a workflow to automatically publish a release to npm once a release was tagged.

This ensures that we publish with provenance and increase trust into the builds.

Read more here: https://docs.npmjs.com/generating-provenance-statements

You now need to tag a commit with `v`+ version string and create a release on GitHub to trigger the publish workflow.